+++ src/main/java/fr/maxlego08/menu/loader/actions/CurrencyWithdrawLoader.java
@@ -19,6 +19,7 @@ public class CurrencyWithdrawLoader extends ActionLoader {
         String bigDecimal = accessor.getString("amount");
         Currencies currencies = Currencies.valueOf(accessor.getString("currency", Currencies.VAULT.name()).toUpperCase());
         String economyName = accessor.getString("economy", null);
-        return new CurrencyWithdrawAction(bigDecimal, currencies, economyName);
+        String reason = accessor.getString("reason", "no reason");
+        return new CurrencyWithdrawAction(bigDecimal, currencies, economyName, reason);
     }
 }