+++ src/main/java/fr/maxlego08/menu/requirement/actions/CurrencyWithdrawAction.java
@@ -13,15 +13,17 @@ public class CurrencyWithdrawAction extends ActionHelper {
     private final String amount;
     private final Currencies currencies;
     private final String economyName;
+    private final String reason;
 
-    public CurrencyWithdrawAction(String amount, Currencies currencies, String economyName) {
+    public CurrencyWithdrawAction(String amount, Currencies currencies, String economyName, String reason) {
         this.amount = amount;
         this.currencies = currencies;
         this.economyName = economyName;
+        this.reason = reason;
     }
 
     @Override
     protected void execute(Player player, Button button, InventoryEngine inventory, Placeholders placeholders) {
-        this.currencies.withdraw(player, new BigDecimal(papi(placeholders.parse(this.amount), player)), this.economyName == null ? "default" : this.economyName);
+        this.currencies.withdraw(player, new BigDecimal(papi(placeholders.parse(this.amount), player)), this.economyName == null ? "default" : this.economyName, papi(placeholders.parse(this.reason), player));
     }
 }